requirements.txt
openai
langchain
streamlit
langchain-community
pandas
python-dotenv
unidecode
tabulate
chromadb
langchain-openai
tiktoken